+++ test/clients/DummyCommandableHttpClient_test.go
@@ -5,8 +5,6 @@ import (
 
 	cconf "github.com/pip-services3-go/pip-services3-commons-go/config"
 	cref "github.com/pip-services3-go/pip-services3-commons-go/refer"
-	tlogic "github.com/pip-services3-go/pip-services3-rpc-go/test/logic"
-	testservices "github.com/pip-services3-go/pip-services3-rpc-go/test/services"
 )
 
 func TestDummyCommandableHttpClient(t *testing.T) {
@@ -14,30 +12,11 @@ func TestDummyCommandableHttpClient(t *testing.T) {
 	restConfig := cconf.NewConfigParamsFromTuples(
 		"connection.protocol", "http",
 		"connection.host", "localhost",
-		"connection.port", "3000",
+		"connection.port", DummyCommandableHttpServicePort,
 	)
 
-	var service *testservices.DummyCommandableHttpService
-	var client *DummyCommandableHttpClient
-
-	var fixture *DummyClientFixture
-
-	ctrl := tlogic.NewDummyController()
-
-	service = testservices.NewDummyCommandableHttpService()
-	service.Configure(restConfig)
-
-	references := cref.NewReferencesFromTuples(
-		cref.NewDescriptor("pip-services-dummies", "controller", "default", "default", "1.0"), ctrl,
-		cref.NewDescriptor("pip-services-dummies", "service", "http", "default", "1.0"), service,
-	)
-	service.SetReferences(references)
-
-	service.Open("")
-	defer service.Close("")
-
-	client = NewDummyCommandableHttpClient()
-	fixture = NewDummyClientFixture(client)
+	client := NewDummyCommandableHttpClient()
+	fixture := NewDummyClientFixture(client)
 
 	client.Configure(restConfig)
 	client.SetReferences(cref.NewEmptyReferences())