+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Remote Procedure Calls for Pip.Services in Go Changelog
 
+## <a name="1.3.2"></a> 1.3.2 (2021-05-06)
+
+### Features
+* **test** Refactor test services running
+* Encode URL params
+
 ## <a name="1.3.1"></a> 1.3.1 (2021-04-23) 
 
 ### Features